refactor(Race): 경주의 결과와 출력에 관한 메서드들을 Race.js 파일로 분리
